Modularized controller
