+++ Application/server/routes/authRouter.js
@@ -1,5 +1,4 @@
 const express = require('express');
-const userController = require('../controllers/userController');
 const router = express.Router();
 const passport = require('passport');
 const googlePassport = require('../authenticate');
