+++ Application/server/server.js
@@ -2,15 +2,13 @@ const express = require('express');
 const cors = require('cors');
 const userController = require('./controllers/userController.js');
 const session = require('express-session');
-// import cookieParser from 'cookie-parser';
-// import http from 'http';
 const dotenv = require('dotenv');
 dotenv.config();
-// Make sure server is connected to mongoDB database
 const connectDB = require('./db.js');
 const port = process.env.PORT || 3000;
 const app = express();
 connectDB();
+
 
 // Server connects to the frontend without CORS restriction
 app .use(cors({
@@ -27,24 +25,22 @@ app .use(cors({
         saveUninitialized: true,
       })
     )
-    // .use(cookieParser());
+
 
 const userRouter = require('./routes/userRouter');
 const taskRouter = require('./routes/taskRouter');
 const userinfoRouter = require('./routes/userinfoRouter');
 const authRouter = require('./routes/authRouter');
+const projectRouter = require('./routes/projectRouter')
+
 
 // ROUTES
 app .use('/tasks', taskRouter)
     .use('/userinfo', userinfoRouter)
     .use('/user', userRouter)
     .use('/auth', authRouter)
+    .use('/project', projectRouter)
 
-
-// Route for creating a new project
-app.post('/create/project', userController.createProject, (req, res) => {
-  return res.status(200).json(res.locals.currProject);
-});
 
 // app.get('/dashboard', (req, res) => {
 //   return res.status(200).json();
